Do Metrics folder creation logic during initial load

+++ server/routes/api.js
@@ -33,7 +33,6 @@ router.get('/m/metrics/:id',
 // GET to '/report/:html'
 //  - serves the html file to the page from reports folder
 router.get('/report/:title', (req, res) => {
-  console.log(__dirname);
   //res.status(200).sendFile(`./Metrics/Desktop/Lighthouse/${req.params.title}.html`);
   res.status(200).sendFile(path.join(__dirname, `../../Metrics/Desktop/Lighthouse/${req.params.title}.html`));
 });
@@ -45,19 +44,30 @@ router.get('/m/report/:title', (req, res) => {
 // GET to '/'
 //  - Clear all screenshots and lighthouse reports in desktop and mobile folders
 router.get('/clear', async (req, res) => {
-  for (const file of await fs.readdirSync('Metrics/Desktop/Screenshots')) {
-    await fs.unlinkSync(`Metrics/Desktop/Screenshots/${file}`);
+  if (!fs.existsSync('Metrics')) {
+    fs.mkdirSync('Metrics');
+    fs.mkdirSync('Metrics/Desktop');
+    fs.mkdirSync('Metrics/Desktop/Screenshots');
+    fs.mkdirSync('Metrics/Desktop/Lighthouse');
+    fs.mkdirSync('Metrics/Mobile');
+    fs.mkdirSync('Metrics/Mobile/Screenshots');
+    fs.mkdirSync('Metrics/Mobile/Lighthouse');
+    res.status(200).send('Created all folders');
+  } else {
+    for (const file of await fs.readdirSync('Metrics/Desktop/Screenshots')) {
+      await fs.unlinkSync(`Metrics/Desktop/Screenshots/${file}`);
+    }
+    for (const file of await fs.readdirSync('Metrics/Desktop/Lighthouse')) {
+      await fs.unlinkSync(`Metrics/Desktop/Lighthouse/${file}`);
+    }
+    for (const file of await fs.readdirSync('Metrics/Mobile/Screenshots')) {
+      await fs.unlinkSync(`Metrics/Mobile/Screenshots/${file}`);
+    }
+    for (const file of await fs.readdirSync('Metrics/Mobile/Lighthouse')) {
+      await fs.unlinkSync(`Metrics/Mobile/Lighthouse/${file}`);
+    }
+    res.status(200).send('Cleared all screenshots and reports');
   }
-  for (const file of await fs.readdirSync('Metrics/Desktop/Lighthouse')) {
-    await fs.unlinkSync(`Metrics/Desktop/Lighthouse/${file}`);
-  }
-  for (const file of await fs.readdirSync('Metrics/Mobile/Screenshots')) {
-    await fs.unlinkSync(`Metrics/Mobile/Screenshots/${file}`);
-  }
-  for (const file of await fs.readdirSync('Metrics/Mobile/Lighthouse')) {
-    await fs.unlinkSync(`Metrics/Mobile/Lighthouse/${file}`);
-  }
-  res.status(200).send('Cleared all screenshots and reports');
 });
 
 // POST to '/addURL'
